Overall changes. Start of structure refactoring
