--- src/com/kupybaton/Product.java
@@ -1,25 +1,0 @@
-package com.kupybaton;
-
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-
-public class Product {
-	private Integer id;
-	private String name;
-
-	public Integer getId() {
-		return id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public Product(Integer id, String name) {
-		this.id = id;
-		this.name = name;
-
-	}
-
-}
