+++ src/com/kupybaton/web/servlets/ListSelect.java
@@ -1,9 +1,15 @@
-package com.kupybaton;
+package com.kupybaton.web.servlets;
 
-import javax.servlet.*;
-import javax.servlet.http.*;
-import java.io.*;
-import java.util.*;
+import com.kupybaton.model.Product;
+import com.kupybaton.web.jdbc.ListsExpert;
+
+import javax.servlet.RequestDispatcher;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.List;
 
 public class ListSelect extends HttpServlet {
 	public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
