+++ webapps/Kupybaton.html
@@ -17,7 +17,7 @@ and open the template in the editor.
 <h1> List of products </h1>
 </center>
 
-<p><input type="text" name="listname" value = "name of the list">  Name of the list</p>
+<p><input type="text" name="listname" placeholder="name of the list">  Name of the list</p>
 
 <p><input type="checkbox" name="product" value = "1, 'black_bread'">black_bread</p>
